requirements.txt
# ============================================
# CHOCODEALERS BOT DEPENDENCIES
# Python 3.11+
# ============================================

# Telegram Bot
python-telegram-bot==20.8
python-telegram-bot[job-queue]==20.8

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Square API
square==35.1.0.20240320

# Google APIs
google-api-python-client==2.115.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==6.0.0
gspread-asyncio==1.9.1
oauth2client==4.1.3

# HTTP & API
aiohttp==3.9.3
requests==2.31.0

# Configuration
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Date & Time
python-dateutil==2.8.2
pytz==2024.1

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.40.3

# Data Processing
pandas==2.2.0
openpyxl==3.1.2

# Utils
pyyaml==6.0.1
python-json-logger==2.0.7

# Development & Testing
pytest==8.0.0
pytest-asyncio==0.23.4
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Optional: AI/Analytics (for Notebook LM integration)
openai==1.12.0
anthropic==0.18.1
